Potential fix for code scanning alert no. 39: Multiplication result converted to larger type
To fix the problem, we need to ensure that the multiplication of width and height is performed using the larger type size_t to prevent overflow. This can be achieved by casting one of the operands to size_t before performing the multiplication. This way, the multiplication will be done using size_t, and the result will be correctly assigned to the data_size variable.

+++ libwebp-1.0.3/src/enc/alpha_enc.c
@@ -298,7 +298,7 @@ static int EncodeAlpha(VP8Encoder* const enc,
   const int height = pic->height;
 
   uint8_t* quant_alpha = NULL;
-  const size_t data_size = width * height;
+  const size_t data_size = (size_t)width * height;
   uint64_t sse = 0;
   int ok = 1;
   const int reduce_levels = (quality < 100);
